requirements_dev.txt
ipywidgets==8.0.4
notebook==6.5.2
ensure==1.0.2
opencv-python
numpy==1.21.0
imutils==0.5.4
types-waitress
types-requests
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0
waitress==2.1.2
flask_cors==3.0.10
##
#Dev requirements
pytest==7.2.1
tox==4.4.6
black==23.1.0
# flake8==6.0.0
mypy==1.0.1
pillow==9.4.0
requests==2.28.2

-e .